implemented forcefield gpu kernel
for performance, depending on the number of atom pairs in the energy function, I'm seeing anywhere from a 1x to a 45x speedup! Looks promising so far! =D

had to rework the gpu kernel api a bit to accomodate the new fancier kernel

I haven't implemented the reduction (adding all the energies together) on the gpu yet. I'll do that next. It's a bit tricky to get right.