fix bottleneck for gpu-based minimization!! =)
turns out the protein geometry update is super fast, but the CCD implementation uses decomposed energy functions for each DOF. Which means, at each step in the minimization, CCD is only evaluating a small subset of the energy function, which is way faster than evaluating the whole thing.

so I implemented dof decomposition for gpu forcefields too and that got rid of the bottleneck. The trouble is, since CCD spends most of its time evaluating a smaller energy function, the gpu speedup isn't as large as it is on the full energy function. nevertheless, this commit raises the gpu speedup for full minimization to roughly the empirical max measured by BenchmarkForcefieldKernel (but for efuncs the size of intra+shell rather than full conf).

the bad news is the speedups for intra+shell sized energy functions aren't as large as for full conf energy functions. the good news is, the gpu speedups will actually scale with protein size! The larger the protein, the faster the gpu is than the cpu! =D

also, I took the opportunity to refactor energy function interfaces here to reduce dependencies between modules (eg EPIC and MoleculeModifierAndScorer)