add simpler pfunc framework with implementation
the existing pfunc implementations do lots and lots of wicked awesome
stuff, but are very complicated. I wanted to make a very specialized
pfunc implementation that focuses as much as possible on only
calculating partition functions in the hopes of making that process as
fast as humanly possible. partition functions get calculated a lot, so
speedups here will pay off very well everywhere else they're used.

my implementation has as few "external" dependencies as possible so it
can be easily used in a wide variety of different environments,
like in unit tests for example. =) it's completely fire-and-forget.
meaning, once you configure it, it has all the info it needs to compute
the partition function and it doesn't need to go asking around in global
state (or near-global state) for more info. it also tries to outsource
all the complicated configuration-dependent stuff (like how to compute
conformation energies and where do A* trees come from) to other reusable
objects in the same way the new GMECFinder does.

my implementation is currently single-threaded, but because it uses the
new reusable components like ConfEnergyCalculator, adding support for
multiple threads and even GPUs should be very easy. I'll do that next.

I think this implementation is using the latest and greatest ideas about
how to bound sums of decreasing boltzmann weights too, so it should
require a minimal number of conformation minimizations, like the
ParallelN implementations do. In single-threaded tests, it seems to
perform about as well as the ParallelN implementations, and it's
certainly much much better than the Traditional implementation.

I still need to add support for checkpointing/resuming. I'll work on
that next too