rename pfunc impl from "simple" to "parallelConf"
simple was such a stupid name anyway...

as far as I can tell, the existing parallel implementations compute
q* approximations and q' approximations (and p* approximations) in
different threads

the new parallelConf implementation does the parallelism differently.
it uses the new parallel minimization tools to evaluate multiple
conformations in parallel, rather than the values of the pfunc in
parallel

for multiple threads, this approach is much faster. it also allows us
to use multiple gpus. when using all 8 gpus on the fancy CS dept
machines, K* is super duper fast!