implement a quasi-Newton minimizer
brand new minimizer implemented from scratch

the energies aren't quite as low as the CCD minimizer, but it's pretty
damn close!

this serial cpu implementation is quite a bit slower than the CCD
implementation though, so there's not much use for this minimizer as-is

this algorithm also has much more inherent parallelism though, so the
eventual gpu implementation should be super fast! I'm estimating maybe
2-3 times faster than our CCD + gpu energy implementation? won't know
for sure until I write the gpu kernel