optimize BigForcefieldEnergy a lot, and other speedups
the CCD kernel is stupid fast with enough concurrency, but we were
spending too much time updating the energy function after chemical
changes to the molecule (eg mutations).

chemical updates are fast now, but the new bottleneck is creating and
destroying minimizers. doing lots of parallel memory malloc/free on the
gpu seems to be pretty slow. the next project will be to reuse buffers
between minimizations somehow. we need some kind of minimizer pool, but
that will require some big changes to the conf minimizer