move runtime files to a single resources dir
and putting the resources dir on the classpath lets us load the files
at runtime using Java's resource system, which works even when all the
files are built into a single jar file

this will eventually let us eliminate some unnecessary Osprey config,
like path to the data files, etc. If all the data files are on the
classpath, we'll always know where they are.