started removing the environment vars
first, I started replacing calls to PDBFileReader with calls to
PDBIO instead. PDBFileReader is deprecated now and will be removed
in a future commit.

second, I added a Strand class that deals with picking out contiguous
subsets of molecules and doing bond assignment and template matching.
It needs to know about forcefields and template libraries, but if you
don't explicitly specify one, it will look to the Defaults class for
default forcefields, template libraries, etc.

There's a huge comment at the top of Defaults for how (in my view)
configuration should be handled in Osprey. Take a look if you're
interested in programming philosophy at all.

also fixed a bunch of bugs in PDBIO

also re-did the templates-from-alternates code to use the new PDBIO

I think I broke COMETS though ;_;  the unit test is failing because
of an empty A* tree. Need to figure out why next time.

At least all the other tests pass. =)