fix stuff I broke in COMETS
sadly, it looks like I broke this near the beginning of the build branch
=(

but the problem went unnoticed for a long time because the unit test for
COMETS still passed! it seems that energy matrices were being read from
the filesystem rather than being re-calculated, so an error in the
forcefield params (caused by improper config handling on my part) caused
the re-calculated energy matrices to be wrong, but we couldn't tell
until we explicitly re-calculated them.

I've fixed the config issue in COMETSDoer

and just to be extra sure this doesn't happen again, I've removed the
cached energy matrices from the repo, and updated the unit test to
delete them automatically.