WIP: implement simpler energy matrix calculator
and do a crap ton of testing to make sure it works

fun fact: this doesn't work yet, but there's a unit test that reports
the failure. I still need to commit this though so I can fix a bug on
another branch, hence the Work In Progress (WIP) prefix.

happy bug day!