add javadoc parsing capabilities to the python docs
so now we can auto-copy javadoc comments into the python docs!

that way we won't have to duplicate documentation for the python API.
we just write the docs once for the java code, and expose it in the python
docs. Or write the docs directly in the python code (ie for real python
functions/classes/etc) and use special tags to pull snippets from the
javadocs if we need it.