convert from Jerkar build system to Gradle+Kotlin, and redo python dists
Gradle is a widely popular build system for JVM apps, but its reliance on Groovy
as a scripting language made it a huge pain in the ass to work with. Much more trouble
than it was worth in my opinion. So I went with the much simpler and plain Java build tool
Jerkar initially. Now Gradle has a Kotlin scripting option that's a complete game-changer.

I ported all of the old Jerkar functionality to Gradle, and also added a bunch of new tasks
to handle the python build processes too. The python build task now deals with the JPype
dependency by using pre-compiled binaries rather than trying to install the source dist from
PyPI. Hopefully this should work much better for end users that don't happen to have a C++
compiler installed.