Add pdb writing and fixing memory issues
Now we clear some stored values that we don't need from the partition function calculation as we go and we write out PDB files for 10 conformations (if they improve stability). We also only check if the windows overlap if we have at least 10 conformations. This prevents us from only having one conformation to rely on.