Do not load the same conflib twice
This fixes a regression introduced with storing the conflibs "by mol",
where the same L-conflib is loaded twice, but when writing the conflib
it is not able to match up conformations with the correct conflib. This
prevented normal protein:protein or protein:peptide designs (whoops, have
been doing too many reflected chirality designs recently).

Remove the unit test that reproduced the issue, because this fix avoids the
issue by only loading the built-in conflibs once and saves a per-molecule
pointer to the conflibs.