update GPU codes for nvidia toolkit 12.3
Reduce the threads in launch bound to 512 so that it will compile.
Otherwise it uses too many threads per streaming multiprocessor. Should
probably now adjust the number of concurrent blocks in the launch
bounds.

add two colons prior to std:: objects as the compiler mistakenly thought
we wanted to refer to cuda::std, which doesn't exist.