Add min repro test for NativeConfEnergyCalc failure
Calculation of an energy matrix using the NativeConfEnergyCalculator
fails when one of the bodies has translation and rotation enabled. This
PR adds unit tests with minimal working examples to reproduce the error.
The example is a conformation space created from PDB 6ov7 (CALP:kCAL01)
with no mutations on either molecule, no amino acid flexibility on
either molecule, and translation+rotation on kCAL01.