Bugfix: trans+rot fixed in NativeConfEnergyCalc
The number of possible modified positions is now set to the number of
molecular motions (eg translation + rotations) + the number of design
positions. Prior to this change, designs that used the
NativeConfEnergyCalculator (C++ calculator) encountered an
"index-out-of-bounds" assertion error when one of the molecules had
translation/rotation enabled.

TestMakeEnergyMatrixWithNativeCalculators.java verifies that
EnergyMatrices can be constructed by all types of supported calculators.
Prior to this change, construction of energy matrices with the
NativeConfEnergyCalculator failed.